requirements.txt
flask>=3.0.0
pyyaml>=6.0
pydantic>=2.0
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
mcp>=1.0.0
